add missing files

+++ v2/Rpc/Bench.Server/Worker/Counters/Counter.cs
@@ -23,7 +23,7 @@ namespace Bench.RpcSlave.Worker.Counters
             ResetCounters();
         }
 
-        public void PushAll(Action<string, int> handler)
+        public void PushAll(Action<string, double> handler)
         {
             lock (InnerCounters)
             {
